Add support for SAF
Use the normal (fast) file interface when possible, but add
support for SAF when necessary.

Also introduces an introduction to improve the first usage, but
it is also necessary for SAF.

Finally, it removes the automatic backup, as DecSync should
take care of that. Manual importing and exporting OPML is
still supported.